+++ lib/generators/cms/component/search_page/templates/app/cells/search_cell.rb
@@ -1,0 +1,8 @@
+class SearchCell < Cell::Rails
+  def show(page)
+    @query = params[:q]
+    @search_page = page.homepage.search_page
+
+    render
+  end
+end